+++ projekt_koda/src/main/java/com/example/repository/KomentarjiRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KomentarjiRepository extends JpaRepository<Komentarji, Long> {
+public interface KomentarjiRepository extends JpaRepository<Komentarji, Integer> {
 }
