+++ projekt_koda/src/main/java/com/example/repository/PrehranskeOmejitveRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PrehranskeOmejitveRepository extends JpaRepository<PrehranskeOmejitve, Long> {
+public interface PrehranskeOmejitveRepository extends JpaRepository<PrehranskeOmejitve, Integer> {
 }
