+++ projekt_koda/src/main/java/com/example/repository/ReceptSestavineRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReceptSestavineRepository extends JpaRepository<ReceptSestavine, Long> {
+public interface ReceptSestavineRepository extends JpaRepository<ReceptSestavine, Integer> {
 }
